backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.0.1
opencv-python==4.8.1.78
numpy==1.24.3
torch==2.1.0
torchvision==0.16.0
tensorflow==2.14.0
mediapipe==0.10.7
rembg==2.0.50
ultralytics==8.0.196
scikit-learn==1.3.0
matplotlib==3.7.2
requests==2.31.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.21
alembic==1.12.0
pydantic==2.4.2
python-dotenv==1.0.0
boto3==1.29.7
redis==5.0.1